requirements.txt
Flask==0.10.1
numpy==1.14.1+mkl
requests==2.18.4
Pillow==5.0.0
